+++ packages/runtime/src/index.ts
@@ -42,6 +42,7 @@ import { routerHistory as history } from './history.js';
 import KeepAliveOutlet from './KeepAliveOutlet.js';
 import ClientOnly from './ClientOnly.js';
 import useMounted from './useMounted.js';
+import { withSuspense, useSuspenseData } from './Suspense.js';
 
 export {
   getAppConfig,
@@ -81,6 +82,9 @@ export {
   AppErrorBoundary,
   ClientOnly,
   useMounted,
+
+  withSuspense,
+  useSuspenseData,
 };
 
 export type {